+++ game/dataStructures/path.js
@@ -1,0 +1,12 @@
+export default function Path(startPoint, endPoint) {
+  var me = this;
+  me.end = endPoint;
+  me.start = startPoint;
+
+  me.next = function() {
+    return null;
+  };
+  me.prev = function() {
+    return null;
+  };
+}
