+++ game/main.js
@@ -1,41 +1,69 @@
-"use strict";
+'use strict';
+import sharedData from './sharedData.js';
 
-BOXED_GAME.gameLoop = (function(game) {
-	var GameState = game.gameStates.GameState;
-	var stack = game.variables.stateStack;
-	var pauseState = game.gameStates.pauseState;
+(function main() {
+  window.onkeydown = function(event) {
+    sharedData.variables.keyMap[event.keyCode] = event.type == 'keydown';
+  };
 
-	function updateTimeStep() {
-		game.variables.now = performance.now();
-		game.variables.dt = game.variables.now - (game.variables.lastUpdate || game.variables.now);
-		game.variables.lastUpdate = game.variables.now;
-	}
+  window.onkeyup = function(event) {
+    sharedData.variables.keyMap[event.keyCode] = event.type == 'keydown';
+  };
 
-	game.variables.getCurrentGameState().load();
-	// "Game loop" this is where the continous function goes 
-	function tick() {
+  window.onblur = function() {
+    var game = BOXED_GAME;
+    var pauseScreen = game.gameStates.pauseScreen;
+    var currentState = sharedData.variables.getCurrentGameState();
 
-		updateTimeStep();
-		game.constants.CONTEXT2D.clearRect(0, 0, game.constants.CANVAS_WIDTH, game.constants.CANVAS_HEIGHT);
-		var currentState = game.variables.getCurrentGameState();
+    if (!sharedData.variables.isPaused &&
+        sharedData.constants.STATE_TYPES.get('action') === currentState.type) {
+      pauseScreen.load();
+      sharedData.variables.isPaused = true;
+      sharedData.variables.stateStack.push(pauseScreen);
+    }
+  };
 
-		currentState.control();
-		currentState.update();
-		currentState.draw();
-		if ( !currentState.isPlaying ) {
-			game.constants.CONTEXT2D.clearRect(0, 0, game.constants.CANVAS_WIDTH, game.constants.CANVAS_HEIGHT);
+  var requestAniFrame = window.requestAnimationFrame ||
+      window.webkitRequestAnimationFrame || window.mozRequestAnimationFrame ||
+      window.oRequestAnimationFrame ||
+      window.msRequestAnimationFrame || function(callback) {
+        window.setTimeout(callback, 1000 / sharedData.constants.FPS_LIMIT);
+      };
 
-			if ( game.variables.stateStack.length > 0 ) {
-				game.variables.stateStack.pop();
-				currentState = game.variables.getCurrentGameState();
-				currentState.load();
-			} else {
-				return;
-			}
-		}
-		requestAniFrame(tick);
-	}
-	requestAniFrame(tick);
+  function updateTimeStep() {
+    sharedData.variables.now = window.performance.now();
+    sharedData.variables.dt =
+        game.variables.now - (game.variables.lastUpdate || game.variables.now);
+    sharedData.variables.lastUpdate = game.variables.now;
+  }
 
-}(BOXED_GAME));
+  sharedData.variables.getCurrentGameState().load();
 
+  function tick() {
+    updateTimeStep();
+    sharedData.constants.CONTEXT2D.clearRect(
+        0, 0, sharedData.constants.CANVAS_WIDTH,
+        sharedData.constants.CANVAS_HEIGHT);
+    var currentState = sharedData.variables.getCurrentGameState();
+
+    currentState.control();
+    currentState.update();
+    currentState.draw();
+
+    if (!currentState.isPlaying) {
+      sharedData.constants.CONTEXT2D.clearRect(
+          0, 0, sharedData.constants.CANVAS_WIDTH,
+          sharedData.constants.CANVAS_HEIGHT);
+
+      if (sharedData.variables.stateStack.length > 0) {
+        sharedData.variables.stateStack.pop();
+        currentState = sharedData.variables.getCurrentGameState();
+        currentState.load();
+      } else {
+        return;
+      }
+    }
+    requestAniFrame(tick);
+  }
+  requestAniFrame(tick);
+})();
